read description, especially artists
Added shine and goals files for KR icons that Der Kaiser put in the modding hoi4 discord for anyone to use. These aren't the actual icons as that would take up a lot of space and we aren't going to use all of them just a few. So if you want to see them or want to add some just let me know and I can send a zip of all the KR icons